Feature/dashboard

+++ app/Register/page.js
@@ -18,7 +18,24 @@ export default function Register() {
         e.preventDefault();
         setError('');
 
+        // Client-side validation for password length
+        if (password.length < 6) {
+            setError('Password must be at least 6 characters long.');
+            setErrorColor('text-red-500');
+            setTimeout(() => setError(''), 3000); // Clear error after 3 seconds
+            return;
+        }
+
         try {
+            // Check if email already exists in the database
+            const emailCheckResponse = await axios.post('/api/checkEmailExists', { email });
+            if (emailCheckResponse.data.exists) {
+                setError('An account with this email already exists.');
+                setErrorColor('text-red-500');
+                setTimeout(() => setError(''), 3000); // Clear error after 3 seconds
+                return;
+            }
+
             // Store the password in session storage
             sessionStorage.setItem('password', password);
 
@@ -30,10 +47,12 @@ export default function Register() {
             } else {
                 setError('Failed to send verification email. Please try again.');
                 setErrorColor('text-red-500');
+                setTimeout(() => setError(''), 3000); // Clear error after 3 seconds
             }
         } catch (err) {
             setError('Error creating account');
             setErrorColor('text-red-500');
+            setTimeout(() => setError(''), 3000); // Clear error after 3 seconds
         }
     };
 
@@ -57,7 +76,7 @@ export default function Register() {
             <img
               alt=""
               src="" /* TODO: Add Logo */
-              className="mx-auto h-10 w-auto"
+              className="mx-auto h-24 w-auto"
             />
           </div>
           <div 
@@ -70,17 +89,13 @@ export default function Register() {
             boxShadow: '0px 0px 10px 5px rgba(20,13,1,1)', 
             WebkitBoxShadow: '0px 0px 10px 5px rgba(20,13,1,1)', 
             MozBoxShadow: '0px 0px 10px 5px rgba(20,13,1,1)',
-            border: '2px solid white'
+            border: '2px solid #1E1412',
+            backgroundColor: '#1E1412'
           }}>
-            {error && <p className={`${errorColor} text-right mt-0 pt-0`}>{error}</p>}
             <h2 className="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-white" style={{paddingTop: '0', textShadow: '2px 2px 4px rgba(0, 0, 0, 1)'}}>
               Create Account
             </h2>
-            <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm" 
-              style={{
-                position: 'fixed', 
-                transform: `translate(10%, ${error ? '-12%' : '0%'})`
-              }}>
+            <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm" >
               <form onSubmit={handleSubmit} className="space-y-6">
 
                 <div>
@@ -96,8 +111,8 @@ export default function Register() {
                       autoComplete="firstname"
                       value={firstname}
                       onChange={(e) => setFirstname(e.target.value)}
-                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-[#b79994] sm:text-sm sm:leading-6"
-                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px' }}
+                      className="block w-full bg-[#1E1412] rounded-md border-0 py-1.5 text-gray-300 shadow-sm border-[1px] focus:border-[#C69635] focus:outline-none"
+                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px', boxShadow: '0px 0px 10px 0px rgba(0, 0, 0, 0.1)' }}
                     />
                   </div>
                 </div>
@@ -115,8 +130,8 @@ export default function Register() {
                       autoComplete="lastname"
                       value={lastname}
                       onChange={(e) => setLastname(e.target.value)}
-                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-[#b79994] sm:text-sm sm:leading-6"
-                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px' }}
+                      className="block w-full bg-[#1E1412] rounded-md border-0 py-1.5 text-gray-300 shadow-sm border-[1px] focus:border-[#C69635] focus:outline-none"
+                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px', boxShadow: '0px 0px 10px 0px rgba(0, 0, 0, 0.1)' }}
                     />
                   </div>
                 </div>
@@ -134,8 +149,8 @@ export default function Register() {
                       autoComplete="email"
                       value={email}
                       onChange={(e) => setEmail(e.target.value)}
-                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-[#b79994] sm:text-sm sm:leading-6"
-                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px' }}
+                      className="block w-full bg-[#1E1412] rounded-md border-0 py-1.5 text-gray-300 shadow-sm border-[1px] focus:border-[#C69635] focus:outline-none"
+                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px', boxShadow: '0px 0px 10px 0px rgba(0, 0, 0, 0.1)' }}
                     />
                   </div>
                 </div>
@@ -155,8 +170,8 @@ export default function Register() {
                       autoComplete="current-password"
                       value={password}
                       onChange={(e) => setPassword(e.target.value)}
-                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-[#b79994] sm:text-sm sm:leading-6"
-                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px' }}
+                      className="block w-full bg-[#1E1412] rounded-md border-0 py-1.5 text-gray-300 shadow-sm border-[1px] focus:border-[#C69635] focus:outline-none"
+                      style={{ fontSize: '1rem', fontWeight: 'bold', paddingLeft: '10px', boxShadow: '0px 0px 10px 0px rgba(0, 0, 0, 0.1)' }}
                     />
                   </div>
                 </div>
@@ -172,16 +187,17 @@ export default function Register() {
                 >
                   <button
                     type="submit"
-                    className="flex w-[205px] justify-center rounded-md bg-[#683F24] px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-[#442718] focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-[#b79994]"
-                    style={{ transition: 'background-color 0.3s ease-in-out' }}
+                    className="flex w-[205px] justify-center rounded-md bg-[#1E1412] px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm border border-gray-400 hover:border-[#C69635] rounded"
+                    style={{ transition: 'background-color 0.3s ease-in-out', boxShadow: '0px 0px 10px 0px rgba(0, 0, 0, 0.1)' }}
                   >
                     Create Account
                   </button>
 
                   <a href="/Login"
-                    className="flex w-full justify-center rounded-md bg-[#683F24] px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-[#442718] focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-[#b79994]"
+                    className="flex w-full justify-center rounded-md bg-[#1E1412] px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm border border-gray-400 hover:border-[#C69635] rounded"
                     style={{
                       transition: 'background-color 0.3s ease-in-out',
+                      boxShadow: '0px 0px 10px 0px rgba(0, 0, 0, 0.1)'
                     }}
                   >
                     Have an account? Login here
@@ -202,6 +218,11 @@ export default function Register() {
               className="relative left-[calc(50%+3rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 bg-gradient-to-tr from-[#5A3A2F] to-[#2B1D1A] opacity-30 sm:left-[calc(50%+36rem)] sm:w-[72.1875rem]"
             />
           </div>
+          {error && (
+            <div className={`fixed bottom-4 right-4 bg-red-500 text-white p-4 rounded shadow-lg transition-opacity duration-300 ease-in-out ${error ? 'opacity-100' : 'opacity-0'}`}>
+              {error}
+            </div>
+          )}
         </div>
       </>
     )
